requirements.txt
livekit
livekit-api
numpy
python-dotenv
opencv-python